+++ openaq/_async/models/countries.py
@@ -1,7 +1,14 @@
 from openaq.shared.models import build_query_params
 from openaq.shared.responses import CountriesResponse
-from openaq.shared.utils import validate_integer_id
-
+from openaq.shared.types import SortOrder
+from openaq.shared.validators import (
+    validate_integer_id,
+    validate_integer_or_list_integer_params,
+    validate_limit_param,
+    validate_order_by,
+    validate_page_param,
+    validate_sort_order,
+)
 from .base import AsyncResourceBase
 
 
@@ -42,26 +49,17 @@ class Countries(AsyncResourceBase):
         page: int = 1,
         limit: int = 1000,
         order_by: str | None = None,
-        sort_order: str | None = None,
-        parameters_id: int | None = None,
-        providers_id: int | None = None,
+        sort_order: SortOrder | None = None,
+        parameters_id: int | list[int] | None = None,
+        providers_id: int | list[int] | None = None,
     ) -> CountriesResponse:
         """List countries based on provided filters.
 
-        Provides the ability to filter the countries resource by the given arguments.
-
-        * `page` - Specifies the page number of results to retrieve
-        * `limit` - Sets the number of results generated per page
-        * `providers_id` - Filter results by selected providers ID(s)
-        * `parameters_id` - Filters results by selected parameters ID(s)
-        * `order_by` - Determines the fields by which results are sorted; available values are `id`
-        * `sort_order` - Works in tandem with `order_by` to specify the direction: either `asc` (ascending) or `desc` (descending)
-
         Args:
-            page: The page number. Page count is countries found / limit.
-            limit: The number of results returned per page.
+            page: The page number, must be greater than zero. Page count is countries found / limit.
+            limit: The number of results returned per page. Must be between 1 and 1,000.
             order_by: Order by operators for results.
-            sort_order: Sort order (asc/desc).
+            sort_order: Order for sorting results (asc/desc).
             parameters_id: Single parameters ID or an array of IDs.
             providers_id: Single providers ID or an array of IDs.
 
@@ -69,6 +67,7 @@ class Countries(AsyncResourceBase):
             CountriesResponse: An instance representing the list of retrieved countries.
 
         Raises:
+            InvalidParameterError: Client validation error, query parameter is not correct type or value.
             IdentifierOutOfBoundsError: Client validation error, identifier outside support int32 range.
             ApiKeyMissingError: Authentication error, missing API Key credentials.
             BadRequestError: Raised for HTTP 400 error, indicating a client request error.
@@ -84,6 +83,28 @@ class Countries(AsyncResourceBase):
             ServiceUnavailableError: Raised for HTTP 503, indicating that the server is not ready to handle the request.
             GatewayTimeoutError: Raised for HTTP 504 error, indicating a gateway timeout.
         """
+        page = validate_page_param(page)
+        limit = validate_limit_param(limit)
+        if sort_order is not None:
+            sort_order = validate_sort_order(sort_order)
+        if order_by is not None:
+            order_by = validate_order_by(order_by)
+        if parameters_id is not None:
+            parameters_id = validate_integer_or_list_integer_params(
+                'parameters_id', parameters_id
+            )
+        if providers_id is not None:
+            providers_id = validate_integer_or_list_integer_params(
+                'providers_id', providers_id
+            )
+        params = build_query_params(
+            page=page,
+            limit=limit,
+            order_by=order_by,
+            sort_order=sort_order,
+            parameters_id=parameters_id,
+            providers_id=providers_id,
+        )
         params = build_query_params(
             page=page,
             limit=limit,