+++ openaq/shared/transport.py
@@ -1,11 +1,13 @@
 """Base class and utlity functions for working with client transport."""
 
+from __future__ import annotations
+
 import logging
 from abc import ABC, abstractmethod
 from http import HTTPStatus
-from typing import Any, Mapping
+from typing import Mapping
 
-from httpx import Response
+import httpx
 
 from openaq.shared.exceptions import (
     BadGatewayError,
@@ -24,56 +26,27 @@ from openaq.shared.exceptions import (
 logger = logging.getLogger("transport")
 
 
-class BaseTransport(ABC):
-    """Base class for client transport classes."""
-
-    @abstractmethod
-    async def send_request(
-        self,
-        method: str,
-        url: str,
-        params: Mapping[str, str],
-        headers: Mapping[str, Any],
-    ):
-        """Sends request using transport. To be overridden in subclass.
-
-        Args:
-            method: HTTP method name
-            url: URL to send requestion to
-            params: query parameters to add to URL
-            headers: HTTP headers to include wiht request
-        """
-        raise NotImplementedError
-
-    @abstractmethod
-    async def close(self):
-        """Closes transport connection. To be overridden in subclass."""
-        raise NotImplementedError
-
-
-def check_response(res: Response) -> Response | None:
+def check_response(res: httpx.Response) -> httpx.Response:
     """Checks the HTTP response of the request.
 
-        Args:
-            res: an httpx.Response object
+    Args:
+        res: an httpx.Response object
 
-        Returns:
-            httpx.Response
+    Returns:
+        httpx.Response
 
-        Raises:
-            IdentifierOutOfBoundsError: Client validation error, identifier outside support int32 range.
-    ApiKeyMissingError: Authentication error, missing API Key credentials.
-            BadRequestError: Raised for HTTP 400 error, indicating a client request error.
-            NotAuthorizedError: Raised for HTTP 401 error, indicating the client is not authorized.
-            ForbiddenError: Raised for HTTP 403 error, indicating the request is forbidden.
-            NotFoundError: Raised for HTTP 404 error, indicating a resource is not found.
-            TimeoutError: Raised for HTTP 408 error, indicating the request has timed out.
-            ValidationError: Raised for HTTP 422 error, indicating invalid request parameters.
-            HTTPRateLimitError: Raised for HTTP 429 error, indicating rate limit exceeded.
-            ServerError: Raised for HTTP 500 error, indicating an internal server error or unexpected server-side issue.
-            BadGatewayError: Raised for HTTP 502, indicating that the gateway or proxy received an invalid response from the upstream server.
-            ServiceUnavailableError: Raised for HTTP 503, indicating that the server is not ready to handle the request.
-            GatewayTimeoutError: Raised for HTTP 504 error, indicating a gateway timeout.
+    Raises:
+        BadRequestError: Raised for HTTP 400 error, indicating a client request error.
+        NotAuthorizedError: Raised for HTTP 401 error, indicating the client is not authorized.
+        ForbiddenError: Raised for HTTP 403 error, indicating the request is forbidden.
+        NotFoundError: Raised for HTTP 404 error, indicating a resource is not found.
+        TimeoutError: Raised for HTTP 408 error, indicating the request has timed out.
+        ValidationError: Raised for HTTP 422 error, indicating invalid request parameters.
+        HTTPRateLimitError: Raised for HTTP 429 error, indicating rate limit exceeded.
+        ServerError: Raised for HTTP 500 error, indicating an internal server error or unexpected server-side issue.
+        BadGatewayError: Raised for HTTP 502, indicating that the gateway or proxy received an invalid response from the upstream server.
+        ServiceUnavailableError: Raised for HTTP 503, indicating that the server is not ready to handle the request.
+        GatewayTimeoutError: Raised for HTTP 504 error, indicating a gateway timeout.
     """
     if res.status_code >= HTTPStatus.OK and res.status_code < HTTPStatus.BAD_REQUEST:
         return res